11.jan up

+++ database/migrations/2020_01_10_160914_add_opinion_id_to_users_table.php
@@ -14,7 +14,7 @@ class AddOpinionIdToUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->integer('opinion_id')->after('password');
+            $table->BIGINT('opinion_id')->after('password');
         });
     }
